fix.socket,tower
소켓연결
