+++ public/src/game.js
@@ -3,6 +3,7 @@ import { Monster } from "./monster.js";
 import { Tower } from "./tower.js";
 import towerData from "../assets/tower.json" with { type: "json" };
 import { TowerControl } from "./towerControl.js";
+import { sendEvent } from "./socket.js";
 
 /* 
   어딘가에 엑세스 토큰이 저장이 안되어 있다면 로그인을 유도하는 코드를 여기에 추가해주세요!
@@ -25,7 +26,7 @@ import { TowerControl } from "./towerControl.js";
 
 */
 
-let socket; // 서버 웹소켓 객체
+
 const canvas = document.getElementById("gameCanvas");
 const ctx = canvas.getContext("2d");
 
@@ -429,6 +430,7 @@ function initGame() {
 } //이게 시작이네.
 
 if (!isInitGame) {
+  sendEvent(2,{timestamp: Date.now()})
   initGame();
 }
 
@@ -442,31 +444,7 @@ Promise.all([
     (img) => new Promise((resolve) => (img.onload = resolve))
   ),
 ]).then(() => {
-  let somewhere = localStorage.getItem("authToken");
 
-  socket = io("http://localhost:8080", {
-    auth: {
-      token: somewhere, // 토큰이 저장된 어딘가에서 가져와야 합니다!
-    },
-  });
-
-  socket.on('response', (data) => {
-    console.log(data);
-  });
-  
-  socket.on('connection', (data) => {
-    console.log('connection: ', data);
-  
-    if(data.message){
-      console.log(data.message);
-    }
-  });
-
-  console.log("a");
-  //서버의 이벤트들을 받는 코드들은 여기다가 쭉 작성해주시면 됩니다!
-  //e.g. serverSocket.on("...", () => {...});
-  //이 때, 상태 동기화 이벤트의 경우에 아래의 코드를 마지막에 넣어주세요! 최초의 상태 동기화 이후에 게임을 초기화해야 하기 때문입니다!
- 
 });
 
 // 타워를 설치할 수 있는지 판별하는 함수
