+++ src/models/tower.model.js
@@ -2,9 +2,6 @@ const towers = {};
 
 
 export const createTower = (uuid) => {
-    
-
-
     towers[uuid] = [];
 }
