+++ src/routes/user.router.js
@@ -67,7 +67,7 @@ router.post('/sign-in', async (req, res) => {
     // 비밀번호가 일치하면 JWT 생성
     const token = jwt.sign(
       {
-        userId: user.userId, // JWT 페이로드에 사용자 키 포함 여기서 유저 id(유저 구분을 위한 uuid) 넣기.
+        userId: user.id, // JWT 페이로드에 사용자 키 포함 여기서 유저 id(유저 구분을 위한 uuid) 넣기.
       },
       process.env.JWT_KEY, // 비밀 키를 사용하여 서명
       { expiresIn: '8h' } // 토큰 유효 기간을 8시간으로 설정
@@ -97,7 +97,7 @@ router.get('/users', async (req, res, naxt) => {
     //여기서 decoded 안에있는 유저 찾아서 비교.
     const user = await prisma.users.findUnique({
       where: {
-        userId: decoded.userId,
+        id: decoded.id,
       },
     });
     if (!user) {
